+++ src/hla/lattice.py
@@ -33,6 +33,7 @@ class Lattice:
         self.chromaticity = [0.0, 0.0]
         self.circumference = 0.0
         self.orm = None
+        self.loop = True
 
     def __getitem__(self, key):
         if isinstance(key, int):
@@ -288,7 +289,7 @@ class Lattice:
             if group in self._group.keys():
                 #print "found exact group", group
                 #print self._group.keys()
-                return self._group[group]
+                return self._group[group][:]
 
             # do pattern match on element name
             ret, names = [], []
@@ -561,30 +562,39 @@ class Lattice:
         """
         Assuming self._elements is in s order
 
-        the element matched with input 'element' string should be unique.
+        the element matched with input 'element' string should be unique
+        and exact.
+
+        If the input *element* name is also in *group*, no duplicate the
+        result. 
+
+        ::
+
+          >>> getNeighbors('P4', 'BPM', 2)
+          ['P2', 'P3', 'P4', 'P5', 'P6']
+          >>> getNeighbors('Q3', 'BPM', 2)
+          ['P2', 'P3', 'Q3', 'P4', 'P5']
         """
 
-        e0 = self.getElements(element)
-        if len(e0) > 1:
-            raise ValueError("element %s is not unique" % element)
-        elif e0 == None or len(e0) == 0:
-            raise ValueError("element %s does not exist" % element)
+        e0 = self._find_element(element)
+        if not e0: raise ValueError("element %s does not exist" % element)
 
-        e, s = self.getElements(group, point = 'e')
-        #print e, s
+        el = self.getElements(group)
 
-        i1, i2 = 0, 0
-        ret = [[element[:], s0]]
-        for i in range(0, len(s)):
-            if s[i] > s0[0]:
-                #return e[i-1], s[i-1], e[i], s[i]
-                i1, i2 = i - 1, i
-                break 
-            #else: print "s", s0[0], s[i]
+        if not el: raise ValueError("elements/group %s does not exist" % group)
+        if e0 in el: el.remove(e0)
+
+        i0 = len(el)
+        for i,e in enumerate(el):
+            if e.sb < e0.sb: continue
+            i0 = i
+            break
+        ret = [e0]
         for i in range(n):
-            ret.insert(0, [e[i1-i], s[i1-i]])
-            ret.append([e[i2+i], s[i2+i]])
-
+            fac, r = divmod(i0 - i - 1, len(el))
+            ret.insert(0, el[r])
+            fac, r = divmod(i0 + i, len(el))
+            ret.append(el[r])
         return ret
         
     def __repr__(self):